dealer/requirements.txt
python-telegram-bot>=21.0
python-dotenv
treys
websockets>=13.0
